rename AppFunctional.js to App.js
